requirements.txt
flask==2.3.3
flask-jwt-extended==4.5.2
